+++ generic/tkbltGraph.h
@@ -239,7 +239,7 @@ namespace Blt {
 
     int getXY(const char*, int*, int*);
     void getTextExtents(Tk_Font, const char*, int, int*, int*);
-    void getBoundingBox(int, int, float, double*, double*, Point2d*);
+    void getBoundingBox(int, int, double, double*, double*, Point2d*);
     Point2d anchorPoint(double, double, double, double, Tk_Anchor);
 
     const char** getTags(ClientData, ClassId, int*);
